refactor: optimize  dataset-related dialog interactions
