+++ apps/web/src/app/dataset/[dataset_id]/settings/documents/[document_id]/segments/delete-segment.tsx
@@ -1,65 +1,51 @@
+import { useModal } from '@ebay/nice-modal-react'
+import { Trash } from 'lucide-react'
 import useSWRMutation from 'swr/mutation'
 
-import {
-  AlertDialog,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from '@/components/ui/alert-dialog'
 import { Button } from '@/components/ui/button'
+import ConfirmDialog from '@/components/nice-confirm-dialog'
 
-import { editSegment } from './segment-form'
+import { editSegment } from './segment-form-dialog'
 import { ISegmentDeleteProps } from './type'
 
 const DeleteSegment = ({
   dataset_id = '',
   uid = '',
   segment_id = '',
-  setShowDeleteAlter,
-  showDeleteAlter,
-  handelConfirm,
+  handelRefresh,
 }: ISegmentDeleteProps) => {
-  const { trigger, isMutating } = useSWRMutation(
-    `/api/datasets/segment`,
-    editSegment
-  )
+  const { trigger } = useSWRMutation(`/api/datasets/segment`, editSegment)
+  const modal = useModal(ConfirmDialog)
 
-  const handelDelete = async () => {
+  const onOk = async () => {
     await trigger({ dataset_id, uid, segment_id, content: '' })
-    setShowDeleteAlter?.(false)
-    handelConfirm?.()
+    handelRefresh()
   }
+
+  const handleDelete = () => {
+    modal.show({
+      title: `Delete Segment?`,
+      description: ` Are you sure you want to delete this segment? This action cannot be
+      undone.`,
+      confirmText: 'Delete',
+      loadingText: 'Deleting...',
+      onOk,
+    })
+  }
+
   return (
-    <AlertDialog onOpenChange={setShowDeleteAlter} open={showDeleteAlter}>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>Delete Segment?</AlertDialogTitle>
-          <AlertDialogDescription>
-            Are you sure you want to delete this segment? This action cannot be
-            undone.
-          </AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <Button
-            variant="outline"
-            type="button"
-            onClick={() => setShowDeleteAlter?.(false)}
-          >
-            Cancel
-          </Button>
-          <Button
-            onClick={handelDelete}
-            variant="destructive"
-            type="button"
-            disabled={isMutating}
-          >
-            {isMutating ? 'Deleting' : 'Delete'}
-          </Button>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
+    <Button
+      type="button"
+      className="invisible absolute bottom-4 right-4 flex h-8 w-8 gap-2 text-red-600 group-hover/card:visible"
+      size="icon"
+      variant="outline"
+      onClick={(e) => {
+        e.stopPropagation()
+        handleDelete()
+      }}
+    >
+      <Trash size={18} />
+    </Button>
   )
 }
 export default DeleteSegment
