+++ apps/web/src/app/dataset/[dataset_id]/settings/documents/[document_id]/segments/type.ts
@@ -2,9 +2,7 @@ export interface ISegmentDeleteProps {
   dataset_id: string
   uid: string
   segment_id: string
-  showDeleteAlter: boolean
-  setShowDeleteAlter: (s: boolean) => void
-  handelConfirm: () => void
+  handelRefresh: () => void
 }
 
 export interface ISegmentEditProps {
@@ -12,7 +10,5 @@ export interface ISegmentEditProps {
   segment_id: string
   dataset_id: string
   document_id: string
-  open: boolean
-  setOpen: (s: boolean) => void
-  handelConfirm: (s: void) => void
+  handelRefresh: (s: void) => void
 }
