+++ apps/web/src/app/dataset/[dataset_id]/settings/documents/[document_id]/segments/view-page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Suspense, useCallback, useMemo, useRef, useState } from 'react'
+import { Suspense, useCallback, useMemo, useState } from 'react'
 import { useSearchParams } from 'next/navigation'
 import { getCoreRowModel, useReactTable } from '@tanstack/react-table'
 import { debounce } from 'lodash'
@@ -9,10 +9,7 @@ import useSWR from 'swr'
 import { fetcher } from '@/lib/utils'
 import { Input } from '@/components/ui/input'
 import { Skeleton } from '@/components/ui/skeleton'
-import { SegmentProps } from '@/app/dataset/type'
 
-import AddOrEdit from './add-edit-segment'
-import DeleteSegment from './delete-segment'
 import SegmentHeader from './header'
 import { DataTablePagination } from './pagination'
 import SegmentList from './segment-list'
@@ -37,8 +34,6 @@ const LoadingCards = () => (
 )
 
 const SegmentPage = ({ datasetId, name, type, icon, appId }: IProps) => {
-  const [open, setOpen] = useState(false)
-  const [showDeleteAlter, setShowDeleteAlter] = useState(false)
   const [value, setValue] = useState('')
   const [fresh, setFresh] = useState(0)
 
@@ -46,7 +41,6 @@ const SegmentPage = ({ datasetId, name, type, icon, appId }: IProps) => {
     pageSize: 100,
     pageIndex: 0,
   })
-  const current = useRef<SegmentProps>({ content: '', segment_id: '' })
 
   async function getDatasetDocument(
     params: [queries: Record<string, any>, pagination: Record<string, any>]
@@ -94,10 +88,7 @@ const SegmentPage = ({ datasetId, name, type, icon, appId }: IProps) => {
           icon={icon}
           type={type}
           appId={appId}
-          addNew={() => {
-            setOpen(true)
-            current.current = { content: '', segment_id: '' }
-          }}
+          handelRefresh={() => setFresh((v) => v + 1)}
         />
       </Suspense>
       <Suspense fallback={<LoadingCards />}>
@@ -116,9 +107,9 @@ const SegmentPage = ({ datasetId, name, type, icon, appId }: IProps) => {
               <div className="grid grid-cols-1 gap-4 md:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4 ">
                 <SegmentList
                   segments={data?.segments}
-                  setOpen={setOpen}
-                  current={current}
-                  setShowDeleteAlter={setShowDeleteAlter}
+                  datasetId={datasetId}
+                  uid={uid}
+                  handelRefresh={() => setFresh((v) => v + 1)}
                 />
               </div>
             )}
@@ -130,24 +121,6 @@ const SegmentPage = ({ datasetId, name, type, icon, appId }: IProps) => {
           )}
         </div>
       </Suspense>
-
-      <DeleteSegment
-        dataset_id={datasetId}
-        uid={uid}
-        segment_id={current?.current?.segment_id || ''}
-        showDeleteAlter={showDeleteAlter}
-        setShowDeleteAlter={setShowDeleteAlter}
-        handelConfirm={() => setFresh((v) => v + 1)}
-      />
-      <AddOrEdit
-        content={current?.current?.content || ''}
-        open={open}
-        segment_id={current?.current?.segment_id || ''}
-        dataset_id={datasetId}
-        document_id={uid}
-        setOpen={setOpen}
-        handelConfirm={() => setFresh((v) => v + 1)}
-      />
     </div>
   )
 }
